% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/australia.R
\name{australia}
\alias{australia}
\title{australia}
\usage{
australia(site)
}
\arguments{
\item{site}{French gauge number}
}
\value{
data frame of discharge time-series
}
\description{
Provides access to Australian gauge data
}
\examples{
df = australia('403213')
plot(df$Date, df$Q, type='l')
}
