% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brazil.R
\name{brazil}
\alias{brazil}
\title{brazil}
\arguments{
\item{site}{Brazilian gauge number}
}
\value{
data frame of discharge time-series
}
\description{
Provides access to Japanese gauge data
}
\examples{
df = brazil('12650000')
plot(df$Date, df$Q, type='l')
}
