% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/france.R
\name{france}
\alias{france}
\title{france}
\arguments{
\item{site}{French gauge number}
}
\value{
data frame of discharge time-series
}
\description{
Provides access to French gauge data
}
\examples{
df = france('K027401001')
plot(df$Date, df$Q, type='l')
}
