% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/australia.R
\name{get_hourly}
\alias{get_hourly}
\title{get_hourly}
\usage{
get_hourly(
  parameter_type,
  station_number,
  start_date,
  end_date,
  tz,
  return_fields
)
}
\arguments{
\item{parameter_type}{Parameter value}

\item{station_number}{Station number}

\item{start_date}{Start date (formatted as YYYY-MM-DD) or just the
year (YYYY)}

\item{end_date}{End date (formatted as YYYY-MM-DD) or just the year (YYYY)}

\item{tz}{TZ}

\item{return_fields}{Return fields}
}
\description{
get_hourly
}
\examples{
# Hourly streamflow Cotter River at Gingera Gauge
\dontrun{
get_hourly(
  parameter_type = "Water Course Discharge",
  station_number = "410730",
  start_date = "2020-01-01",
  end_date = "2020-01-31"
)
}
}
