% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/japan.R
\name{japan}
\alias{japan}
\title{japan}
\arguments{
\item{site}{Japanese gauge number}
}
\value{
data frame of discharge time-series
}
\description{
Provides access to Japanese gauge data
}
\examples{
df = chile('01201005')
plot(df$Date, df$Q, type='l')
}
