% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-methods.R
\name{original}
\alias{original}
\title{Get original data}
\usage{
original(x, ...)
}
\arguments{
\item{x}{Tibble.}

\item{...}{Additional arguments. None implemented.}
}
\value{
list
}
\description{
Get original data
}
