% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/australia.R
\name{parameters}
\alias{parameters}
\alias{parameters()}
\title{Available water parameters}
\usage{
parameters(pars)
}
\arguments{
\item{pars}{Optional: if empty all available parameters will be returned.
Alternatively, a vector of the continuous or discrete parameters can be
requested.}
}
\value{
A vector of parameters.
}
\description{
\code{parameters} returns a vector of parameters that can be retrieved from
Water Data Online.
}
\examples{
parameters()
parameters("continuous")
parameters("discrete")
}
