% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-methods.R
\name{plot.rr_tbl}
\alias{plot.rr_tbl}
\title{Plot values}
\usage{
\method{plot}{rr_tbl}(x, ...)
}
\arguments{
\item{x}{Tibble.}

\item{...}{Additional arguments. None implemented.}
}
\value{
ggplot2
}
\description{
Plot values
}
