% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RivRetrieve-package.R
\docType{package}
\name{RivRetrieve-package}
\alias{RivRetrieve-package}
\title{RivRetrieve: Retrieve global river and stage data}
\description{
Functions to retrieve global river gauge data from various national-level APIs.
}
\examples{
\dontrun{
print("TODO")
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ryan-Riggs/RivRetrieve}
}

}
\author{
\strong{Maintainer}: Ryan Riggs \email{rriggs@tamu.edu} (\href{https://orcid.org/0000-0001-6834-9469}{ORCID})

Authors:
\itemize{
  \item Simon Moulds \email{sim.moulds@gmail.com} (\href{https://orcid.org/0000-0002-7297-482X}{ORCID})
}

}
