% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{uk}
\alias{uk}
\title{uk}
\arguments{
\item{site}{UK gauge number}
}
\value{
data frame of discharge time-series
}
\description{
Provides access to UK gauge data
}
\examples{
df = uk("http://environment.data.gov.uk/hydrology/id/stations/3c5cba29-2321-4289-a1fd-c355e135f4cb")
plot(df$Date, df$Q, type='l')
}
