% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada.R
\name{canada}
\alias{canada}
\title{canada}
\usage{
canada(site)
}
\arguments{
\item{site}{Canadian gauge number}
}
\value{
data frame of discharge time-series
}
\description{
Provides access to Canadian gauge data
}
\examples{
df = canada("01AD003")
plot(df$Date, df$Q, type='l')
}
