% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chile.R
\name{chile}
\alias{chile}
\title{chile}
\arguments{
\item{site}{Chilean gauge number}
}
\value{
data frame of discharge time-series
}
\description{
Provides access to Chile gauge data
}
\examples{
df = chile('01201005')
plot(df$Date, df$Q, type='l')
}
