% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chile.R
\name{chile}
\alias{chile}
\title{chile}
\usage{
chile(site, variable = "discharge", start_date = NULL, end_date = NULL, ...)
}
\arguments{
\item{site}{Chilean gauge number}

\item{variable}{Character. Either \code{stage} or \code{discharge}.}

\item{start_date}{Character. Optional start date with format
YYYY-MM-DD. Default is 1900-01-01.}

\item{end_date}{Character. End date with format YYYY-MM-DD.
Default is the current date.}

\item{...}{Additional arguments. None implemented.}
}
\value{
data frame of discharge time-series
}
\description{
Retrieve Chilean gauge data
}
\examples{
\dontrun{
df <- chile('01201005')
plot(df$Date, df$Q, type='l')
}
}
