% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/australia.R
\name{get_monthly}
\alias{get_monthly}
\title{get_monthly}
\usage{
get_monthly(
  parameter_type,
  station_number,
  start_date,
  end_date,
  tz,
  return_fields
)
}
\description{
get_monthly
}
\examples{
# Monthly average dry air temperature at Corin Dam
\dontrun{
get_monthly(
  parameter_type = "Dry Air Temperature",
  station_number = "570947",
  start_date = "2016-01-01",
  end_date = "2016-06-01"
)
}
}
