% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RivRetrieve-package.R
\docType{package}
\name{RivRetrieve-package}
\alias{RivRetrieve-package}
\title{RivRetrieve: Retrieve global river and stage data}
\description{
RivRetrieve provides access to global river gauge data from a variety of national-level river agencies. The package interfaces with the national-level agency websites to provide access to river gauge locations, river discharge, and river stage. Currently, the package is available for the following countries: Australia, Brazil, Canada, Chile, France, Japan, South Africa, the United Kingdom, and the United States.
}
\examples{
\dontrun{
print("TODO")
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ryan-Riggs/RivRetrieve}
}

}
\author{
\strong{Maintainer}: Ryan Riggs \email{rriggs@tamu.edu} (\href{https://orcid.org/0000-0001-6834-9469}{ORCID})

Authors:
\itemize{
  \item Simon Moulds \email{sim.moulds@gmail.com} (\href{https://orcid.org/0000-0002-7297-482X}{ORCID})
  \item Michel Wortmann \email{michel.wortmann@ouce.ox.ac.uk} (\href{https://orcid.org/0000-0002-1879-7674}{ORCID})
  \item Louise Slater \email{louise.slater@ouce.ox.ac.uk} (\href{https://orcid.org/0000-0001-9416-488X}{ORCID})
  \item George Allen \email{geoallen@vt.edu} (\href{https://orcid.org/0000-0001-8301-5301}{ORCID})
}

}
