% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southAfrica.R
\docType{data}
\name{southAfrica}
\alias{southAfrica}
\alias{station_specific}
\title{southAfrica}
\format{
An object of class \code{character} of length 1.
}
\usage{
station_specific
}
\arguments{
\item{site}{South African gauge number}
}
\value{
data frame of discharge time-series
}
\description{
Provides access to French gauge data
}
\examples{
df = southAfrica('X3H023')
plot(df$Date, df$Q, type='l')
}
\keyword{datasets}
