% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southAfrica.R
\name{southAfrica}
\alias{southAfrica}
\alias{seperateFun}
\title{southAfrica}
\usage{
seperateFun(x)
}
\arguments{
\item{site}{South African gauge number}
}
\value{
data frame of discharge time-series
}
\description{
Provides access to French gauge data
}
\examples{
df = southAfrica('X3H023')
plot(df$Date, df$Q, type='l')
}
