% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa.R
\name{usa}
\alias{usa}
\alias{remove_has_internet}
\title{usa}
\usage{
remove_has_internet()
}
\arguments{
\item{site}{US gauge number}
}
\value{
data frame of discharge time-series
}
\description{
Provides access to USA gauge data
}
\examples{
df = usa("02471078")
plot(df$Date, df$Q, type='l')
}
